services/data_pipeline/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
newspaper3k==0.2.8
pyyaml==6.0.1
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-huggingface==0.0.1
chromadb==0.4.22
sentence-transformers==2.6.0
python-dotenv==1.0.0
lxml[html_clean]
